corrected a bug in run-on words due to last latter being a diacritic
added support for run-on compund words.

added names used by eNarrator to our dictionary

refined hadith state-machine to support for breaking when we have ibn or abi followed by punctuation to break tolerence, was not supported before

corrected cases to have rasoul words accept prefixes by parsing them inside stemmer

added support for numbers (digits) breaking tolerance

corrected bug in display of start and end of hadith (used to be 1 letter off and not the last word but the one after it)

added support for not recomputing whether a word is rasoul or not but have it saved in the narrator once known.