corrected bug in equal_ignoring_diacritics similar to that previously corrected in equal
corrected bug due to possessive place names (namePrim startpos not initialized) anyways. seems possessive places are instead of adding to accuracy hurting it.

also implemented learning (i.e. detected previosuly unknown names, but we dont mean totally learning since they are not used later in parsing just detected as names, nor added to dictionary:

-after عن must come a narrator, but this is not treated as a name
-after 2ikhbar which has not been followed by any NMC, a word that has no valid morphological analysis is treated as Name. a word that has a morphological analysis that makes it NOUN_PROP (i.e. from those of buckwalter, our names added but not used eg: female names, those of eNarrator (introduces many false positives) )
-after 2ikhbar followed even by a valid word followed by a punctuation treated as narrator.

we also supported for breaking the sanad if a fullstop along with a newline are encountered.

we supported for sanads not starting except if by ikhbar (anywhere in the sentence), or 3an which starts a sentence (not in a middle of sentence), and a name at the start of a sentence... all other forms are invalid.

now not using names of eNarrator in our name detection except if for learned names in the cases documented above, bc adding much false positives (و, بن, له, etc... all treated as names)